requirements.txt
pillow==11.2.1
opencv-python==4.11.0.86
numpy==2.3.0